moved caching of config to separate class in order to write the right data

+++ app/Console/Commands/StresstestSetup.php
@@ -7,6 +7,8 @@ use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\DB;
 use Symfony\Component\Process\Process;
 
+
+use Illuminate\Contracts\Console\Kernel as ConsoleKernelContract;
 class StresstestSetup extends Command
 {
     /**
@@ -42,21 +44,27 @@ class StresstestSetup extends Command
     {
         $envFile = '.env';
         $envBackupFileWhileStresstest = ".envBackupWhileStresstest";
+        if(file_exists($envBackupFileWhileStresstest)) {
+            print_r([
+                config('app.env'),
+            ]);
+            exit;
+        }
+        if (config('app.env') !== 'local') {
 
-        if (env('APP_ENV') !== 'local') {
             if(!file_exists($envBackupFileWhileStresstest)) {
                 $this->error('You cannot perform this action on this environment! only with APP_ENV set to local AND not in production (read config:cache && route:cache)!!');
                 return false;
             }
             else{
                 if($this->confirm('You didn\'t do a teardown first, do you want to prepare for another stresstest?')){
-                    $this->error('Stresstest setup was cancelled by you');
-                    return false;
-                }
-                else{
                     $this->info('we\'re going to do a teardown first');
                     Artisan::call(sprintf('stresstest:teardown'));
                     $this->info('teardown done');
+                }
+                else{
+                    $this->error('Stresstest setup was cancelled by you');
+                    return false;
                 }
             }
         }
@@ -70,7 +78,7 @@ class StresstestSetup extends Command
         }
 
         $sqlImports = [
-            'streetestdb.sql',
+            'stresstestdb.sql',
         ];
 
         $this->info('start refreshing database...(this can take some time as in several minutes)');
@@ -83,26 +91,35 @@ class StresstestSetup extends Command
 
         $this->info(PHP_EOL);
 
-        $this->info('going to set env settings to production');
-        $this->printSubItem('make backup of '.$envFile.' to '.$envBackupFileWhileStresstest);
-        $envContents = file_get_contents($envFile);
-        $this->info('done');
-        $this->printSubItem('set app_env  to production and debug to false');
-        file_put_contents($envBackupFileWhileStresstest,$envContents);
-        $envContents = str_replace('APP_ENV=local','APP_ENV=production',$envContents);
-        $envContents = str_replace('APP_DEBUG=true','APP_DEBUG=false',$envContents);
-        $envContents = str_replace('QUEUE_DRIVER=sync','QUEUE_DRIVER=database',$envContents);
-        file_put_contents($envFile,$envContents);
-        $this->info('done');
-
-        $this->info('Start caching');
-        foreach(['route','config','view'] as $type){
-            $this->printSubItem($type.' cache');
-            Artisan::call(sprintf('%s:clear',$type));
-            Artisan::call(sprintf('%s:cache',$type));
+        if(!file_exists($envBackupFileWhileStresstest)) {
+            $this->info('going to set env settings to production');
+            $this->printSubItem('make backup of ' . $envFile . ' to ' . $envBackupFileWhileStresstest);
+            $envContents = file_get_contents($envFile);
             $this->info('done');
-        };
-        $this->info('caching complete');
+            $this->printSubItem('set app_env  to production and debug to false');
+
+            file_put_contents($envBackupFileWhileStresstest, $envContents);
+
+            $envContents = str_replace('APP_ENV=local', 'APP_ENV=production', $envContents);
+            $envContents = str_replace('APP_DEBUG=true', 'APP_DEBUG=false', $envContents);
+            $envContents = str_replace('QUEUE_DRIVER=sync', 'QUEUE_DRIVER=database', $envContents);
+
+            file_put_contents($envFile, $envContents);
+
+            $this->info('done');
+        }
+
+        /**
+         * added to separate file and needs to be called independently php artisan stresstest:setup && php artisan stresstest:cache
+         */
+//        $this->info('Start caching');
+
+//        foreach(['config','route','view'] as $type){
+//            $this->printSubItem($type.' cache');
+//            Artisan::call(sprintf('%s:cache',$type));
+//            $this->info('done');
+//        };
+//        $this->info('caching complete');
 
         $this->info(PHP_EOL);
         $this->info(PHP_EOL);
@@ -121,7 +138,9 @@ class StresstestSetup extends Command
     }
 
     protected function handleSqlFiles($sqlImports = []){
+        $path = base_path('database/seeds/');
         foreach ($sqlImports as $file) {
+            $file = sprintf('%s%s',$path,$file);
             if(!file_exists($file)){
                 $this->error('The file '.$file.' doesn\'t seem to exist, we can\'t do a proper setup');
                 return false;
@@ -129,8 +148,9 @@ class StresstestSetup extends Command
         }
         foreach ($sqlImports as $file) {
             $this->printSubItem(sprintf('importing %s...',$file));
+            $file = sprintf('%s%s',$path,$file);
             $command = sprintf(
-                'mysql -h %s -u %s -p%s %s < database/seeds/%s',
+                'mysql -h %s -u %s -p%s %s < %s',
                 DB::connection()->getConfig('host'),
                 DB::connection()->getConfig('username'),
                 DB::connection()->getConfig('password'),
